+++ packages/component/src/Activity/CarouselFilmStrip.js
@@ -128,7 +128,6 @@ const WebChatCarouselFilmStrip = ({
 
   return (
     <div
-      aria-label={ariaLabel}
       className={classNames(
         ROOT_CSS + '',
         carouselFilmStripStyleSet + '',
@@ -140,8 +139,9 @@ const WebChatCarouselFilmStrip = ({
         direction === 'rtl' ? 'webchat__carousel--rtl' : ''
       )}
       ref={scrollableRef}
-      role="region"
+      role="group"
     >
+      <ScreenReaderText text={ariaLabel} />
       {renderAvatar && <div className="webchat__carouselFilmStrip__avatar">{renderAvatar()}</div>}
       <div className="content">
         {!!activityDisplayText && (
@@ -155,7 +155,7 @@ const WebChatCarouselFilmStrip = ({
                 }
               })}
             </Bubble>
-            <div className="filler" />
+            <div aria-hidden={true} className="filler" />
           </div>
         )}
         <ul className={classNames({ webchat__carousel__item_indented: indented })} ref={itemContainerRef}>